# api/docs/swagger.yaml
basePath: /
definitions:
  routes.AgentCheckinData:
    properties:
      cmd_result:
        type: string
      os:
        type: string
    type: object
  routes.AgentParsed:
    properties:
      checked:
        type: string
      cmd_result:
        type: string
      id:
        type: string
      name:
        type: string
      os:
        type: string
    type: object
  routes.NoteContent:
    properties:
      content:
        type: string
    type: object
  routes.RunCommandData:
    properties:
      command:
        type: string
      directory:
        type: string
    type: object
  routes.TokenResponse:
    properties:
      token:
        example: eyJhbGci...
        type: string
    type: object
  routes.UserLogin:
    properties:
      name:
        example: test
        type: string
      password:
        example: test
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A sample API
  title: orbitalC2 api
  version: "1.0"
paths:
  /api/v1/agent/agent/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.AgentCheckinData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Agent checkin
      tags:
      - agent
  /api/v1/agent/command/{name}:
    post:
      consumes:
      - application/json
      parameters:
      - description: command data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.RunCommandData'
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Command agent
      tags:
      - user
  /api/v1/agent/create/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new C2 agent
      tags:
      - user
  /api/v1/agent/delete/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Agent Name
        in: path
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete table
      tags:
      - user
  /api/v1/agent/get_command:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.RunCommandData'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: get command
      tags:
      - agent
  /api/v1/agent/list:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/routes.AgentParsed'
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: list agents
      tags:
      - user
  /api/v1/agent/list/inactive:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/routes.AgentParsed'
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: list inactive agents
      tags:
      - user
  /api/v1/notes/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List notes on an agent
      tags:
      - agent notes
  /api/v1/notes/{name}/{noteName}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      - description: Note Name
        in: path
        name: noteName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: get note from agent name and note name, returns base64
      tags:
      - agent notes
  /api/v1/notes/create/{name}/{noteName}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      - description: Note Name
        in: path
        name: noteName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: create note from note name and agent name.
      tags:
      - agent notes
  /api/v1/notes/delete/{name}/{noteName}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      - description: Note Name
        in: path
        name: noteName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: delete note from note name and agent name.
      tags:
      - agent notes
  /api/v1/notes/update/{name}/{noteName}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      - description: Note Name
        in: path
        name: noteName
        required: true
        type: string
      - description: note content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.NoteContent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: update note from note name and agent name.
      tags:
      - agent notes
  /api/v1/user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.TokenResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login for API access
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"